Добавил паттерн "Команда"
