+++ src/main/java/com/fin/spr/command/DataLoaderCommand.java
@@ -1,0 +1,5 @@
+package com.fin.spr.command;
+
+public interface DataLoaderCommand {
+    void execute();
+}
