+++ src/main/java/com/fin/spr/services/KudaGoSerivce.java
@@ -1,6 +1,9 @@
 package com.fin.spr.services;
 
 import com.fin.spr.annotations.LogExecutionTime;
+import com.fin.spr.command.DataLoaderInvoker;
+import com.fin.spr.command.commands.CategoryDataLoaderCommand;
+import com.fin.spr.command.commands.LocationDataLoaderCommand;
 import com.fin.spr.controllers.payload.LocationPayload;
 import com.fin.spr.models.Category;
 import com.fin.spr.models.Event;
@@ -8,6 +11,7 @@ import com.fin.spr.models.Location;
 import com.fin.spr.models.response.EventResponse;
 import com.fin.spr.models.response.EventsResponse;
 import com.fin.spr.storage.InMemoryStorage;
+import lombok.extern.slf4j.Slf4j;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Qualifier;
@@ -40,48 +44,19 @@ import java.util.concurrent.TimeUnit;
  */
 
 @Service
+@Slf4j
 public class KudaGoSerivce  {
-    private static final Logger logger = LoggerFactory.getLogger(KudaGoSerivce.class);
-    private static final String Fields = "title,dates,price,is_free,location";
-
-    @Value("${kudago.api.categories.url}")
-    private String CATEGORIES_API_URL;
-
-    @Value("${kudago.api.locations.url}")
-    private String LOCATIONS_API_URL;
-
-    @Value("${kudago.api.events.url}")
-    private String EVENTS_API_URL;
 
     @Value("${app.init.schedule-duration}")
     private Duration scheduleDuration;
 
-    @Value("${kudago.limits.events.page_size}")
-    private int PAGE_SIZE;
-
-    @Value("${kudago.limits.events.max_page}")
-    private int MAX_PAGE;
-
-    private final RestClient restClient;
-    private InMemoryStorage<Category, Integer> categoryStorage;
-    private final ExecutorService fixedThreadPool;
     private final ScheduledExecutorService scheduledThreadPool;
-    private final LocationService locationService;
-    private final EventService eventService;
+    private final DataLoaderInvoker dataLoaderInvoker;
 
-
-    public KudaGoSerivce(RestClient restClient,
-                         InMemoryStorage<Category, Integer> categoryStorage,
-                         LocationService locationService,
-                         EventService eventService,
-                         @Qualifier("fixedThreadPool") ExecutorService fixedThreadPool,
-                         @Qualifier("scheduledThreadPool") ScheduledExecutorService scheduledThreadPool) {
-        this.restClient = restClient;
-        this.categoryStorage = categoryStorage;
-        this.locationService = locationService;
-        this.eventService = eventService;
-        this.fixedThreadPool = fixedThreadPool;
+    public KudaGoSerivce(@Qualifier("scheduledThreadPool") ScheduledExecutorService scheduledThreadPool,
+                         DataLoaderInvoker dataLoaderInvoker) {
         this.scheduledThreadPool = scheduledThreadPool;
+        this.dataLoaderInvoker = dataLoaderInvoker;
     }
 
     /**
@@ -107,97 +82,6 @@ public class KudaGoSerivce  {
      * </p>
      */
     public void parallelInitData() {
-        try {
-            fixedThreadPool.invokeAll(List.of(
-                    () -> { loadCategories(); return null; },
-                    () -> { loadLocations(); return null; } ,
-                    () -> { loadEvents(); return null; }
-            ));
-            logger.info("Data initialization completed with invokeAll.");
-        } catch (InterruptedException e) {
-            logger.error("Data initialization interrupted.", e);
-            Thread.currentThread().interrupt();
-        }
-    }
-
-    /**
-     * Loads categories from the KudaGo API and stores them in {@link InMemoryStorage}.
-     * <p>
-     * If no data is returned, a warning message is logged.
-     * </p>
-     */
-    private void loadCategories() {
-        List<Category> categories = restClient.get()
-                .uri(CATEGORIES_API_URL)
-                .retrieve()
-                .body(new ParameterizedTypeReference<>() {});
-        if (categories != null) {
-            categories.forEach(category -> categoryStorage.create(category.getId(), category));
-            logger.info("Categories have been successfully initialized.");
-        } else {
-            logger.warn("Received null categories.");
-        }
-    }
-
-    private void loadLocations() {
-        List<LocationPayload> locations = restClient.get()
-                .uri(LOCATIONS_API_URL)
-                .retrieve()
-                .body(new ParameterizedTypeReference<>() {});
-        if (locations != null) {
-            locationService.initLocations(locations);
-            logger.info("Locations have been successfully initialized.");
-        } else {
-            logger.warn("Received null locations.");
-        }
-    }
-
-    private void loadEvents() {
-        List<EventsResponse> allEventsResponse = fetchAllEvents();
-        logger.info("Get " + allEventsResponse.size() + " events pages");
-        if (!allEventsResponse.isEmpty()) {
-            List<EventResponse> eventsToAdd = new ArrayList<>();
-            allEventsResponse.forEach(eventsResponse -> eventsToAdd.addAll(eventsResponse.getResults()));
-            eventService.initEvents(eventsToAdd);
-            logger.info("Events have been successfully initialized.");
-        } else {
-            logger.warn("Received null events.");
-        }
-    }
-
-    private List<EventsResponse> fetchAllEvents(){
-        int page = 1;
-        List<EventsResponse> eventsResponses = new ArrayList<>();
-        while (page <= MAX_PAGE) {
-            var events = fetchEvents(page++);
-            if (events == null) break;
-            eventsResponses.add(events);
-        }
-        return eventsResponses;
-    }
-    private EventsResponse fetchEvents(int page) {
-        logger.info("Get events from page " + page);
-        try {
-            var response = restClient.get()
-                    .uri(uriBuilder -> uriBuilder
-                            .path(EVENTS_API_URL)
-                            .queryParam("page", page)
-                            .queryParam("page_size", PAGE_SIZE)
-                            .queryParam("fields", Fields)
-                            .build())
-                    .retrieve()
-                    .toEntity(EventsResponse.class);
-            if (!(response.getStatusCode().is2xxSuccessful() && response.getBody() != null)) {
-                logger.info("Get nothing from page" + page);
-                return null;
-            }
-            logger.info("Get successfull from page" + page);
-            logger.info("Cnt elements = " + response.getBody().getResults().size());
-            return response.getBody();
-        } catch (Exception ex) {
-            logger.info("(ex) Get nothing from page" + page);
-            logger.error(ex.getMessage());
-            return null;
-        }
+        dataLoaderInvoker.executeCommands();
     }
 }
